Tests/reorganize: move utf8CharBoundary test to own file
